archived/streamlit_demo/requirements.txt
boto3==1.26.73
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
torch>=2.6.0
torchvision>=0.17.0
